Do you mean change it to this?
I'm attempting to change the code on your forked copy of "Starting-Out-With-Python-Programming-Challenges". By attempting to change the source code on your forked copy, I am allowed to propose this change, but not changed it altogether, since it's your copy. Please to Try the same on my master copy of "Starting-Out-With-Python-Programming-Challenges" so that I can direct student to the updated copy.